* hjhwang/customer:
  디자이너 상세 조회 모달 작업중

# Conflicts:
#	Customer/build.gradle
#	Customer/src/main/java/com/h2/dogrooming/admin/Admin.java
#	Customer/src/main/java/com/h2/dogrooming/admin/AdminController.java
#	Customer/src/main/java/com/h2/dogrooming/admin/AdminRepository.java
#	Customer/src/main/java/com/h2/dogrooming/admin/AdminService.java
#	Customer/src/main/java/com/h2/dogrooming/designer/Designer.java
#	Customer/src/main/java/com/h2/dogrooming/designer/DesignerAddress.java
#	Customer/src/main/java/com/h2/dogrooming/reservation/Reservation.java
#	Customer/src/main/resources/templates/fragments/header.html
#	Customer/src/main/resources/templates/page/mypage.html
#	Customer/src/main/resources/templates/page/signup.html

+++ Customer/src/main/java/com/h2/dogrooming/product/Product.java
@@ -19,7 +19,7 @@ public class Product {
 
     @Id
     @GeneratedValue
-    private Integer productId;
+    private long productId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "adminNo")
